fix naming for V1 extraction namespace
