+++ tests/Mindee.IntegrationTests/V1/InvoiceSplitterAutoExtractionTest.cs
@@ -1,6 +1,6 @@
 using Mindee.Input;
 using Mindee.Pdf;
-using Mindee.V1.Extraction;
+using Mindee.V1.Image;
 using Mindee.V1.Parsing.Common;
 using Mindee.V1.Product.Invoice;
 using Mindee.V1.Product.InvoiceSplitter;
